% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpms.R
\name{my_tpms}
\alias{my_tpms}
\title{A function to calculate tpms from your raw count matrix}
\usage{
my_tpms(lengths, counts, genenames, samplenames)
}
\arguments{
\item{lengths}{gene lengths}

\item{counts}{raw count dataframe}

\item{genenames}{row names or other gene names gene per row}

\item{samplenames}{column names representing your sample names}
}
\value{
tpm data frame
}
\description{
A function to calculate tpms from your raw count matrix
}
